[precompile] Split Config.hpp
